squelch compiler warnings in socket.c about integer sizes

+++ lib/socket.c
@@ -89,7 +89,7 @@ int rpc_which_events(struct rpc_context *rpc)
 
 static int rpc_write_to_socket(struct rpc_context *rpc)
 {
-	int64_t count;
+	int32_t count;
 
 	if (rpc == NULL) {
 		return -1;
@@ -133,7 +133,7 @@ static int rpc_read_from_socket(struct rpc_context *rpc)
 	int available;
 	int size;
 	int pdu_size;
-	int64_t count;
+	int32_t count;
 
 #if defined(WIN32)
 	if (ioctlsocket(rpc->fd, FIONREAD, &available) != 0) {
